test: move reaper tests into forward_test.go
Per review feedback, keep server_test.go focused and house the
accept-conn reaper tests alongside the code they cover.
